kfuzztest_bridge: add core components
Introduce kfuzztest_bridge - a tool for sending input to a KFuzzTest
harness based on some stream of random data and a textual description of
the expected input format.

In particular, we introduce a simple DSL for describing a basic
region-based structured input like what KFuzzTest accepts. This
includes the addition of:
- A lexer
- A parser
- An encoder

Encoded pseudo-random inputs can be written to a KFuzzTest target's
input file, thus invoking the target.

Signed-off-by: Ethan Graham <ethangraham@google.com>